refactor(mapgen-studio): extract studio engines out of vite.config into src/server/studio (verbatim move, no topology change)
The serialized queue, both operation stores, instance identity, and the five
engine fns move to createStudioEngines({repoRoot}); the StudioServerContext
builder moves to createStudioServerContext. vite.config consumes the one
process instance for both transports — same surface, same behavior. No
effect-orpc in the extracted import graph (config stays node-evaluable).
